refactor: use updated course and learning materials cards
