+++ src/components/sections/challenges/Learning.tsx
@@ -1,7 +1,7 @@
 import Accordion from "@/components/ui/accordion/Accordion";
 import Section from "@/components/sections/communities/_partials/Section";
 import { LearningModuleCard } from "@/components/cards/LearningModule";
-import CourseMaterial from "@/components/cards/course/CourseMaterial";
+import CourseCard from "@/components/cards/course";
 import { Course, LearningModule } from "@/types/course";
 import { Community } from "@/types/community";
 import { useTranslation } from "next-i18next";
@@ -24,7 +24,7 @@ export default function Learning({ courses, learningModules, community }: { cour
             <div className="text-base font-normal text-primary py-6">{t("communities.overview.challenge.learning.title")}</div>
             <div className={`grid grid-cols-1 gap-3 mb-3 ${courses?.length > 1 && "md:grid-cols-2"}`}>
               {courses?.map((course) => (
-                <CourseMaterial
+                <CourseCard
                   key={`learning-card-data-${course.id}`}
                   title={course.name}
                   description={course.description}
