+++ src/components/sections/communities/overview/LearningMaterials.tsx
@@ -1,6 +1,6 @@
 import CourseCard from "@/components/cards/course"
 import { useMultiSelector } from "@/hooks/useTypedSelector"
-import RelatedLearningCard from "@/components/cards/challenge/_partials/RelatedLearning";
+import { LearningModuleCard } from "@/components/cards/LearningModule";
 import { Course, LearningModule } from "@/types/course";
 import { Community } from "@/types/community";
 import Loader from "@/components/ui/Loader";
@@ -26,22 +26,29 @@ export default function LearningMaterialsOverview() {
 
     return (
         <>
-            <div className="md:hidden">
+            <div className="md:hidden my-8">
                 <div className="font-medium text-.5xl leading-snug">{t("communities.overview.learning-materials.title")}</div>
-                <div className="text-sm font-light lg:w-full lg:pr-7 pt-2 mb-6 md:mb-0">{t("communities.overview.learning-materials.description")} </div>
+                <div className="text-sm font-light lg:w-full lg:pr-7 pt-2">{t("communities.overview.learning-materials.description")} </div>
             </div>
-            <div className="grid gap-6">
+            <div className="grid gap-3 md:gap-6 mb-3 md:mb-6">
                 {courses?.map((course) => {
-                    return <CourseCard key={`learning-material-course-${course.id}`} course={course} community={community} />
+                    return <CourseCard
+                        key={`learning-card-data-${course.id}`}
+                        title={course.name}
+                        description={course.description}
+                        link={`/communities/${community.slug}/courses/${course.slug}`}
+                        duration={course.duration}
+                        level={course.level}
+                        learningModulesCount={course.learningModules?.length}
+                    />
                 })}
             </div>
             <div className="grid md:grid-cols-2 gap-3 w-full">
                 {learningModules?.map((module) => {
-                    return <RelatedLearningCard
-                        key={`learning-material-material-${module.id}`}
-                        title={module.title}
-                        description={module.description}
-                        path={`/communities/${community.slug}`} />
+                    return <LearningModuleCard
+                        key={`related-learning-card-${module.id}`}
+                        data={module}
+                    />
                 })}
             </div>
         </>
