+++ src/components/sections/communities/overview/scoreboard/index.tsx
@@ -26,9 +26,9 @@ export default function ScoreboardOverview(): ReactElement {
 
   return (
     <>
-      <div className="md:hidden">
+      <div className="md:hidden my-8">
         <div className="font-medium text-.5xl leading-snug">{t("communities.overview.scoreboard.title")}</div>
-        <div className="text-sm font-light lg:w-full lg:pr-7 pt-2 mb-6 md:mb-0">{t("communities.overview.scoreboard.description")} </div>
+        <div className="text-sm font-light lg:w-full lg:pr-7 pt-2">{t("communities.overview.scoreboard.description")} </div>
       </div>
       {loading ? (
         <div className="h-full w-full grid">
